+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog
 
     * Fixed `transform` with nested list #531
     * Added curl dependency in Dockerfile #532
+    * Introduce spdx_license_expression
+    * Ability to transform spdx license key from spdx_license_expression to
+      license_expression
 
 2023-08-20
     Release 10.0.0